Last update

+++ src/main/java/progkorny/boatrentalweb01/model/RentalEvent.java
@@ -7,20 +7,37 @@ import jakarta.persistence.ManyToOne;
 
 import java.time.LocalDate;
 
+// Az @Entity annotációval megjelöljük, hogy ez az osztály egy adatbázis-entitás (JPA entitás)
 @Entity
 public class RentalEvent {
+
+    // Az esemény egyedi azonosítója (manuálisan kell beállítani, mivel nincs @GeneratedValue)
     @Id
-    private int id;
+    private long id;
+
+    // Több RentalEvent is hivatkozhat egy hajóra (ManyToOne kapcsolat)
     @ManyToOne
-    @JoinColumn(name = "boat_rented_id")
+    @JoinColumn(name = "boat_rented_id") // Az adatbázisban ez a mező tárolja a hivatkozott hajó ID-ját
     private Boat boatRented;
+
+    // Több RentalEvent is tartozhat egy ügyfélhez
     @ManyToOne
-    @JoinColumn(name = "rental_customer_id")
+    @JoinColumn(name = "rental_customer_id") // Külső kulcs a customer(id)-ra
     private Customer rentalCustomer;
+
+    // A kölcsönzés kezdő dátuma
     private LocalDate rentalDate;
+
+    // A visszahozás dátuma
     private LocalDate returnDate;
+
+    // A bérlés teljes költsége
     private double totalCost;
+
+    // Meg van-e zárva a bérlés (pl. visszahozták-e a hajót)
     private boolean isClosed;
+
+    // --- Getterek és Setterek ---
 
     public Customer getRentalCustomer() {
         return rentalCustomer;
@@ -38,7 +55,9 @@ public class RentalEvent {
         this.boatRented = boatRented;
     }
 
+    // --- Konstruktor Builderrel ---
 
+    // Privát konstruktor, amelyet a Builder használ
     private RentalEvent(Builder builder) {
         this.id = builder.id;
         this.boatRented = builder.boatRented;
@@ -49,14 +68,17 @@ public class RentalEvent {
         this.isClosed = builder.isClosed;
     }
 
+    // Üres konstruktor a JPA működéséhez szükséges
     public RentalEvent() { }
 
+    // Statikus metódus, amivel a Builder-t indíthatjuk
     public static Builder builder() {
         return new Builder();
     }
 
+    // --- Belső Builder osztály ---
     public static class Builder {
-        private int id;
+        private long id;
         private Boat boatRented;
         private Customer rentalCustomer;
         private LocalDate rentalDate;
@@ -64,26 +86,26 @@ public class RentalEvent {
         private double totalCost;
         private boolean isClosed;
 
-        public Builder id(int id) {this.id = id; return this; }
+        public Builder id(long id) { this.id = id; return this; }
         public Builder boat(Boat boat) { this.boatRented = boat; return this; }
         public Builder customer(Customer customer) { this.rentalCustomer = customer; return this; }
         public Builder rentalDate(LocalDate rentalDate) { this.rentalDate = rentalDate; return this; }
         public Builder returnDate(LocalDate returnDate) { this.returnDate = returnDate; return this; }
         public Builder totalCost(double totalCost) { this.totalCost = totalCost; return this; }
-        public Builder isClosed(boolean isClosed) { this.isClosed = isClosed;return this; }
+        public Builder isClosed(boolean isClosed) { this.isClosed = isClosed; return this; }
+
         public RentalEvent build() {
             return new RentalEvent(this);
         }
-
     }
 
-    // Getters and Setters
+    // --- Getterek és setterek ---
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -135,6 +157,7 @@ public class RentalEvent {
         this.isClosed = isClosed;
     }
 
+    // --- toString metódus: naplózáshoz, hibakereséshez hasznos ---
     @Override
     public String toString() {
         return "RentalEvent{" +
